+++ src/main/java/com/vms/demo/dto/fuelingJob/FuelJobDTO.java
@@ -19,13 +19,13 @@ public class FuelJobDTO {
     private Long fuelingJobID;
     private Long carID;
     private Long driverID;
-    private int fuelAmount;
+    private float fuelAmount;
     private int fuelCost;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
     private ZonedDateTime dateTime;
     private String stationName;
     private int mileageBefore;
     private int mileageAfter;
-    private int litersBefore;
-    private int litersAfter;
+    private float litersBefore;
+    private float litersAfter;
 }
